refactor: change rest_api to resources for the base app
