+++ resources/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class RestApiConfig(AppConfig):
-    name = "rest_api"
+    name = "resources"
