+++ setup.cfg
@@ -3,4 +3,4 @@ count = False
 ignore = W191, W504, W503
 max-line-length = 160
 statistics = True
-exclude = venv/*, rest_api/migrations/*, users/migrations/*, studyportal/settings.py
+exclude = venv/*, resources/migrations/*, users/migrations/*, studyportal/settings.py
