+++ users/views.py
@@ -1,7 +1,7 @@
 from users.models import User, FileRequest, CourseRequest, Upload, Notifications
-from rest_api.utils import get_size, get_file_details_and_upload, STRUCTURE
+from resources.utils import get_size, get_file_details_and_upload, STRUCTURE
 from studyportal.settings import NEXUS_URL, SECRET_KEY
-from rest_api.serializers import CourseSerializer
+from resources.serializers import CourseSerializer
 from users.signals import notification_handler
 from rest_framework.response import Response
 from users.serializers import UserSerializer
@@ -12,11 +12,11 @@ from users.serializers import (
     CourseRequestSerializer,
     NotificationsSerializer,
 )
-from rest_api.decorators import post_permitted
+from resources.decorators import post_permitted
 from users.serializers import UploadSerializer
 from studyportal.drive.drive import driveinit
 from users.auth import authorize_user
-from rest_api.models import Course
+from resources.models import Course
 import requests
 import random
 import base64
